+++ src/qwip/qpu/qpu.py
@@ -162,10 +162,13 @@ class QPU:
                         **qwip.converter.unstructure(ch)
                     )
 
+            self.config["compilation/compiler"] = dict(
+                __class__=type(self.compiler).__name__
+            )
+
     def update_frames(self):
-        frames = {}
-        local_oscillators = {
-            key: LO_info["frequency"]
+        frames = {
+            key: Frame(LO_info["frequency"])
             for key, LO_info in self.config["hardware/local_oscillators"].items()
         }
 
@@ -173,7 +176,7 @@ class QPU:
             match system:
                 case QuantumSystem(get_frames=_):
                     frames |= qwip.converter.structure(
-                        system.get_frames(local_oscillators),
+                        system.get_frames(),
                         dict[str, Frame],
                     )
                 case _:
@@ -231,11 +234,11 @@ class QPU:
                     case GMMClassification(
                         measurement_key=k, means=m, covariances=c, num_states=s
                     ):
-                        ro_config["classification"][k] = dict(
+                        ro_config[f"registers/{k}/classification"].update(
                             means=m, covariances=c, num_states=s
                         )
                     case IQRotation(angle) if angle:
-                        ro_config[f"classification/{k}/rotation"] = angle
+                        ro_config[f"registers/{k}/classification/rotation"] = angle
 
     @classmethod
     def load_subsystems(cls, config: ConfigFolder) -> dict[Target, QuantumSystem]: