+++ src/qwip/sequencer/waveform.py
@@ -2,15 +2,21 @@ import itertools as it
 from collections import defaultdict
 from functools import lru_cache
 from numbers import Number, Real
-from typing import TYPE_CHECKING, Any, Self, get_args
+from typing import TYPE_CHECKING, Any, Literal, Self
 
 import attrs
+import matplotlib.pyplot as plt
 import numpy as np
 import sympy as sym
 from attrs import field, validators
 from cattr import Converter
 from loguru import logger
+from matplotlib.axes import Axes
+from matplotlib.figure import Figure
+from matplotlib.ticker import EngFormatter
+from matplotlib.transforms import ScaledTranslation
 from scipy.fft import fft, fftfreq, fftshift
+from scipy.signal import convolve
 
 import qwip
 from qwip._cattr import make_attrs_structure_fn, make_attrs_unstructure_fn
@@ -25,6 +31,7 @@ from qwip.sequencer.utils import (
     _variable_substitution,
 )
 from qwip.typing import is_union_type
+from qwip.utils import deprecated
 
 if TYPE_CHECKING:
     from qwip.sequencer.timeline import Timeline
@@ -44,6 +51,10 @@ def register_waveform(cls) -> type:
 @qfrozen
 class Operation:
     name: str = field(metadata=dict(allow_override=False))
+
+    @property
+    def channel(self) -> str:
+        return ""
 
     @property
     def resolved(self) -> bool:
@@ -85,6 +96,11 @@ class Operation:
         """
         return self
 
+    @classmethod
+    @lru_cache
+    def evolvable(cls) -> set[str]:
+        return set((f.name for f in attrs.fields(cls) if f.init))
+
     def resolve(self, **variable_map) -> Self:
         from qwip.sequencer.timeline import Timeline
 
@@ -96,6 +112,9 @@ class Operation:
         to_update = {}
 
         for f in attrs.fields(type(self)):
+            if not f.init:
+                continue
+
             orig = getattr(self, f.name)
 
             match orig:
@@ -117,13 +136,22 @@ class Operation:
                     new.resolve(**variable_map, inplace=True)
                     to_update[f.name] = new
 
+        if not to_update:
+            return self
+
         return attrs.evolve(self, **to_update)
 
+    def assign_channel(self, new_channel, /) -> Self:
+        """Returns a modified waveform with a new channel."""
+        raise NotImplementedError(
+            f"Cannot assign channel for object of type {type(self)}"
+        )
+
     def evolve(self, **updates):
-        # Separate out fields that are also waveforms.
+        # Separate out fields that are also operations.
         groupby = it.groupby(
             attrs.fields(type(self)),
-            key=lambda f: isinstance(getattr(self, f.name), Waveform),
+            key=lambda f: isinstance(getattr(self, f.name), Operation),
         )
 
         field_names = dict(waveform=[], other=[])
@@ -157,11 +185,11 @@ class Operation:
         to_update = dict()
         # First make pass through non-nested attributes
         for name in field_names["other"]:
-            if name in updates:
+            if name in updates and name in self.evolvable():
                 to_update[name] = updates.pop(name)
 
         for name in field_names["waveform"]:
-            if name in updates:
+            if name in updates and name in self.evolvable():
                 to_update[name] = updates[name]
             else:
                 old = getattr(self, name)
@@ -239,7 +267,7 @@ class Waveform(Operation):
             for f in attrs.fields(type(self)):
                 v = getattr(self, f.name)
 
-                if not isinstance(v, Number) and f.name not in ("name", "channels"):
+                if not isinstance(v, Number) and f.name not in ("name", "channel"):
                     variables.add((f.name, v))
 
             text = (
@@ -252,15 +280,52 @@ class Waveform(Operation):
             else:
                 raise e
 
+    def __mul__(self, other) -> Self:
+        return ConvolvedWaveform(a=self, b=other)
+
     def evaluate_timepoints(self, ts: np.ndarray, **kwargs) -> np.ndarray:
         raise NotImplementedError(
             f"Method evaluate_timepoints not defined for {type(self)}!"
         )
 
-    def plot(self): ...
+    def plot(
+        self,
+        *,
+        ts: np.ndarray | None = None,
+        variables: dict[str, float] = {},
+        ax: Axes | None = None,
+        sample_rate: float | None = None,
+        label: str = "",
+        fig_kwargs: dict = {},
+    ) -> Figure:
+        wave = self.resolve(**variables)
 
-    def fft(self, ts, **kwargs) -> tuple[np.ndarray, np.ndarray]:
-        wave = self(ts, **kwargs)
+        if wvars := wave.variables():
+            raise ValueError(
+                f"Cannot plot wave without concrete values for: {", ".join(wvars)}"
+            )
+
+        if ts is None:
+            sample_rate = sample_rate or 101 / wave.width
+            N = int(sample_rate * wave.width)
+            ts = wave.t0 + np.r_[: N + 1] / sample_rate
+
+        if ax is None:
+            fig, ax = plt.subplots(**fig_kwargs)
+        else:
+            fig = ax.get_figure()
+
+        w_t = wave(ts)
+        ax.plot(ts, w_t.real)
+        ax.plot(ts, w_t.imag)
+        ax.xaxis.set_major_formatter(EngFormatter(unit="s"))
+
+        return fig
+
+    def fft(
+        self, ts: np.ndarray, variables: dict[str, float] = {}
+    ) -> tuple[np.ndarray, np.ndarray]:
+        wave = self(ts, **variables)
 
         if len(wave.shape) > 1 and wave.shape[0] == 2:
             wave = 1j * wave[1] + wave[0]
@@ -272,49 +337,36 @@ class Waveform(Operation):
         return fftshift(ks), fftshift(fs)
 
 
-# Custom structuring of waveform channels to account for legacy serialization.
-def structure_channel(value, cls: type):
-    try:
-        return value["name"]
-    except (KeyError, TypeError):
-        ...
-
-    return qwip.converter.structure(value, str)
-
-
-channels_converter = Converter()
-channels_converter.register_structure_hook(str, structure_channel)
-
-
-def _channels_converter(value):
-    """Converter for Waveform channels.
-
-    This is needed for compatibility with legacy `Channel` classes, which were
-    unstructured as a dictionary with a `"name"` parameter.
-    """
-    try:
-        return channels_converter.structure(value, tuple[str, ...])
-    except Exception:
-        ...
-
-    return _TypeConverter(tuple[str, ...])(value)
-
-
 @register_waveform
 @qfrozen
 class TimedWaveform(Waveform):
-    channels: tuple[str, ...] = field(
-        factory=tuple,
-        metadata=dict(allow_override=False),
-        converter=_channels_converter,
-    )
-    width: NumberOrExpression = 0
     t0: NumberOrExpression = 0
+    width: NumberOrExpression = 0
+    channel: str = field(
+        default="",
+        metadata=dict(allow_override=False),
+    )
+
+    @property
+    @deprecated(
+        version="24.6.0",
+        removed="24.8.0",
+        message=(
+            "Waveforms are now restricted to a single channel. Use `Waveform.channel` "
+            "instead."
+        ),
+    )
+    def channels(self) -> tuple[str]:
+        return (self.channel,) if self.channel else tuple()
 
     def evaluate_timepoints(
         self, ts: np.ndarray, width: float, t0: float, **kwargs
     ) -> np.ndarray:
         return np.zeros((len(self.channels), len(ts)))
+
+    def assign_channel(self, new_channel, /) -> Self:
+        """Returns a modified waveform with a new channel."""
+        return self.evolve(channel=new_channel)
 
 
 @register_waveform
@@ -327,24 +379,108 @@ class Delay(TimedWaveform):
 @qfrozen
 class BasicWaveform(TimedWaveform):
     amplitude: NumberOrExpression = 1
+    phase: NumberOrExpression = 0
+
+    @property
+    def complex_amp(self):
+        return self.amplitude * sym.exp(1j * self.phase * sym.pi / 180)
 
 
 @register_waveform
 @qfrozen
 class InfiniteWaveform(BasicWaveform):
-    width: NumberOrExpression = sym.oo
+    width: NumberOrExpression = field(default=sym.oo, init=False)
 
 
 @register_waveform
 @qfrozen
-class Marker(Waveform):
-    @property
-    def channels(self):
-        return tuple()
+class Marker(TimedWaveform):
+    width: NumberOrExpression = field(default=0, init=False)
+
+    def plot(
+        self,
+        *,
+        ts: np.ndarray | None = None,
+        variables: dict[str, float] = {},
+        ax: Axes | None = None,
+        sample_rate: float | None = None,
+        label: str = "",
+        fig_kwargs: dict = {},
+    ) -> Figure:
+        wave = self.resolve(**variables)
+
+        if not isinstance(wave.t0, Real):
+            raise ValueError(f"Cannot plot Marker with t0 = {self.t0}")
+
+        if ax is None:
+            fig, ax = plt.subplots(**fig_kwargs)
+        else:
+            fig = ax.get_figure()
+
+        label = label or f"{wave.name}"
+        offset = ScaledTranslation(10 / 72, 0, fig.dpi_scale_trans)
+        text_transform = ax.get_xaxis_transform() + offset
+
+        ax.axvline(wave.t0)
+        ax.text(wave.t0, 0.9, label, transform=text_transform)
+        ax.autoscale_view()
+        ax.xaxis.set_major_formatter(EngFormatter(unit="s"))
+
+        return fig
+
+
+@qfrozen
+class ConvolvedWaveform(Waveform):
+    a: Waveform
+    b: Waveform = field()
+
+    @b.validator
+    def _validate_operands(self, attribute, value):
+        a = self.a
+        b = value
+        if a.channel and b.channel and a.channel != b.channel:
+            raise ValueError(
+                f"Cannot convolve waveforms on different channels. Got channels "
+                f"'{a.channel}' != '{b.channel}'"
+            )
 
     @property
-    def width(self):
-        return 0
+    def width(self) -> NumberOrExpression:
+        return self.a.width + self.b.width
+
+    @property
+    def t0(self) -> NumberOrExpression:
+        return self.a.t0 + self.b.t0
+
+    @property
+    def phase(self) -> NumberOrExpression:
+        return self.a.phase + self.b.phase
+
+    @property
+    def channel(self) -> str:
+        return self.a.channel or self.b.channel
+
+    def evaluate_timepoints(self, ts: np.ndarray, **kwargs) -> np.ndarray:
+        t0 = kwargs.get("t0", self.t0)
+
+        start = ts[0]
+        (N,) = ts.shape
+
+        t0 = ((t0 - start) / 2).astype(np.float32)
+
+        t_eval = ts - start
+        t_eval = np.r_[-t_eval[::-1], t_eval[1:]]
+
+        a_t = self.a(t_eval, **(kwargs | dict(t0=t0)))
+        b_t = self.b(t_eval, **(kwargs | dict(t0=t0)))
+
+        w_t = convolve(a_t, b_t, mode="full")
+
+        return w_t[2 * (N - 1) : 3 * (N - 1) + 1]
+
+    def assign_channel(self, new_channel, /) -> Self:
+        """Returns a modified waveform with a new channel."""
+        return self.evolve(a_channel=new_channel, b_channel=new_channel)
 
 
 @register_waveform
@@ -372,9 +508,39 @@ class ReadoutMarker(Marker): ...
 @qfrozen
 class DCWaveform(InfiniteWaveform):
     def evaluate_timepoints(
-        self, ts: np.ndarray, amplitude: float, t0: float, **kwargs
+        self, ts: np.ndarray, amplitude: float, phase: float, t0: float, **kwargs
     ) -> np.ndarray:
-        return amplitude * np.ones_like(ts, dtype=np.float32)
+        rphi = amplitude * np.exp(1j * phase * np.pi / 180)
+        wave = rphi * np.ones_like(ts, dtype=np.complex64)
+        return wave
+
+    def plot(
+        self,
+        *,
+        ts: np.ndarray | None = None,
+        variables: dict[str, float] = {},
+        ax: Axes | None = None,
+        sample_rate: float | None = None,
+        label: str = "",
+        fig_kwargs: dict = {},
+    ) -> Figure:
+        wave = self.resolve(**variables)
+
+        if isinstance(wave.amplitude, Real):
+            raise ValueError(
+                f"Cannot plot DCWaveform with amplitude = {self.ampllitude}"
+            )
+
+        if ax is None:
+            fig, ax = plt.subplots(**fig_kwargs)
+        else:
+            fig = ax.get_figure()
+
+        ax.axhline(wave.t0)
+        ax.autoscale_view()
+        ax.xaxis.set_major_formatter(EngFormatter(unit="s"))
+
+        return fig
 
 
 @register_waveform
@@ -382,9 +548,9 @@ class DCWaveform(InfiniteWaveform):
 class CWWaveform(InfiniteWaveform):
     frequency: Frame
     phase: NumberOrExpression = 0
-    offset: NumberOrExpression = field(
-        default=0, converter=lambda v: float(v) if isinstance(v, int) else v
-    )
+    # offset: NumberOrExpression = field(
+    #     default=0, converter=lambda v: float(v) if isinstance(v, int) else v
+    # )
     frame: Frame | None = None
     hardware_modulation: bool = False
 
@@ -394,11 +560,9 @@ class CWWaveform(InfiniteWaveform):
         ts: np.ndarray,
         amplitude: float,
         phase: float,
-        offset: float | complex,
         phase_tracker: PhaseTracker | None = None,
         frames: dict[str, Frame] = {},
         phase_unit: str = None,
-        complex_out: bool = False,
         **kwargs,
     ) -> np.ndarray:
         """Single frequency waveform.
@@ -413,7 +577,7 @@ class CWWaveform(InfiniteWaveform):
             phase: The starting phase of the modulation tone.
             phase_tracker: A phase tracking dictionary mapping modulation channels
                 to a ndarray of times and discrete phase jumps.
-            phase_unit: Eithe r degrees or radians, specifies the phase units.
+            phase_unit: Either degrees or radians, specifies the phase units.
                 Defaults to the value set in `qsettings['units/phase']`.
 
         Returns:
@@ -444,20 +608,9 @@ class CWWaveform(InfiniteWaveform):
         # Add base modulation at the relevant frequency if doing software modulation
         oscillator = 0 if self.hardware_modulation else software_oscillator
         amplitude = 1 if self.hardware_modulation else amplitude
-        wave = (
-            amplitude * np.exp(1j * (oscillator + phis + phase), dtype=np.complex64)
-            + offset
-        )
+        wave = amplitude * np.exp(1j * (oscillator + phis + phase), dtype=np.complex64)
 
-        if complex_out:
-            return wave
-        elif len(self.channels) <= 1:
-            return wave.real
-        elif len(self.channels) == 2:
-            return wave.view(np.float32).reshape(-1, 2).T
-        else:
-            shape = (max(1, len(self.channels)), len(wave))
-            return np.broadcast_to(wave.real, shape)
+        return wave
 
 
 @register_waveform
@@ -467,43 +620,40 @@ class ModulatedWaveform(Waveform):
     modulation: CWWaveform
 
     @property
-    def width(self) -> float | str:
+    def width(self) -> NumberOrExpression:
         return self.envelope.width
 
     @property
-    def t0(self) -> float | str:
+    def t0(self) -> NumberOrExpression:
         return self.envelope.t0
 
     @property
-    def amplitude(self) -> float | str:
+    def amplitude(self) -> NumberOrExpression:
         A_e = self.envelope.amplitude
-        A_f = self.modulation.amplitude
-        try:
-            return A_e * A_f
-        except TypeError:
-            return f"{A_e} * {A_f}"
+        A_m = self.modulation.amplitude
+
+        return A_e * A_m
 
     @property
-    def channels(self) -> tuple[str, ...]:
-        return self.modulation.channels
+    def phase(self) -> NumberOrExpression:
+        phi_e = self.envelope.phase
+        phi_m = self.modulation.phase
+
+        return phi_e + phi_m
+
+    @property
+    def channel(self) -> str:
+        return self.modulation.channel
 
     def evaluate_timepoints(
         self, ts: np.ndarray, complex_out: bool = False, **kwargs
     ) -> np.ndarray:
-        modulation = self.modulation(ts, complex_out=True, **kwargs)
+        modulation = self.modulation(ts, **kwargs)
         envelope = self.envelope(ts, **kwargs)
 
         wave = envelope * modulation
 
-        if complex_out:
-            return wave
-        elif len(self.channels) <= 1:
-            return wave.real
-        elif len(self.channels) == 2:
-            return wave.view(np.float32).reshape(-1, 2).T
-        else:
-            shape = (max(1, len(self.channels)), len(wave))
-            return np.broadcast_to(wave.real, shape)
+        return wave
 
     def update_phase_tracker(
         self,
@@ -523,10 +673,20 @@ class ModulatedWaveform(Waveform):
             for frame, phase in phis.items():
                 phase_tracker.append(frame, PhaseJump(t0, phase))
 
+    def assign_channel(self, new_channel, /) -> Self:
+        """Returns a modified waveform with a new channel."""
+        return self.evolve(modulation_channel=new_channel)
+
 
 @register_waveform
 @qfrozen
 class VirtualZWaveform(Marker):
+    """A Virtual-Z operation.
+
+    Virtual-Z gates are implemented by applying a phase update to a particular reference
+    frame that tracks the qubit phase evolution.
+    """
+
     frame: Frame
     phase: NumberOrExpression = 0
 
@@ -537,6 +697,25 @@ class VirtualZWaveform(Marker):
         frames: dict[str, Frame] = {},
     ) -> None:
         phase_tracker.append(self.frame, PhaseJump(time, self.phase))
+
+    def plot(
+        self,
+        *,
+        ts: np.ndarray | None = None,
+        variables: dict[str, float] = {},
+        ax: Axes | None = None,
+        sample_rate: float | None = None,
+        label: str = "",
+        fig_kwargs: dict = {},
+    ) -> Figure:
+        return super().plot(
+            ts=ts,
+            variables=variables,
+            ax=ax,
+            sample_rate=sample_rate,
+            label=label or f"VZ[{self.frame}, {self.phase}]",
+            fig_kwargs=fig_kwargs,
+        )
 
 
 @register_waveform
@@ -554,7 +733,13 @@ class PhaseResetWaveform(Marker):
 @qfrozen
 class SquareWaveform(BasicWaveform):
     def evaluate_timepoints(
-        self, ts: np.ndarray, width: float, amplitude: float, t0: float, **kwargs
+        self,
+        ts: np.ndarray,
+        width: float,
+        amplitude: float,
+        phase: float,
+        t0: float,
+        **kwargs,
     ) -> np.ndarray:
         """Square waveform.
 
@@ -562,6 +747,7 @@ class SquareWaveform(BasicWaveform):
             ts: Time values at which to evaluate the pulse.
             width: The width of the square pulse.
             amplitude: The amplitude of the pulse.
+            phase: The phase of the pulse.
             t0: The starting time of the pulse.
 
         Returns:
@@ -569,7 +755,8 @@ class SquareWaveform(BasicWaveform):
             times.
         """
         ts = ts - t0
-        wave = amplitude * ((ts > 0) & (ts < width)).astype(np.float32)
+        rphi = amplitude * np.exp(1j * phase * np.pi / 180)
+        wave = rphi * ((ts > 0) & (ts < width)).astype(np.complex64)
 
         return wave
 
@@ -584,6 +771,7 @@ class GaussianWaveform(BasicWaveform):
         ts: np.ndarray,
         width: float,
         amplitude: float,
+        phase: float,
         t0: float,
         cutoff: float,
         **kwargs,
@@ -595,6 +783,7 @@ class GaussianWaveform(BasicWaveform):
             width: The width of the Gaussian pulse, the Gaussian waveform will be
                 truncated such that w(t) == 0 for t < t0 and t > t0 + width.
             amplitude: The amplitude of the pulse.
+            phase: The phase of the pulse.
             t0: The starting time of the pulse.
             cutoff: How many standard deviations to include in the pulse width.
                 This defines the standard deviation as sigma = width / (2 * cutoff).
@@ -605,7 +794,10 @@ class GaussianWaveform(BasicWaveform):
         """
         ts = ts - t0
         sigma = width / (2 * cutoff)
-        wave = amplitude * np.exp(-0.5 * (ts - width / 2) ** 2 / sigma**2)
+        rphi = amplitude * np.exp(1j * phase * np.pi / 180)
+        wave = rphi * np.exp(
+            -0.5 * (ts - width / 2) ** 2 / sigma**2, dtype=np.complex64
+        )
 
         wave[~((0 <= ts) & (ts <= width))] = 0
 
@@ -625,6 +817,7 @@ class CosineRampWaveform(BasicWaveform):
         ts: np.ndarray,
         width: float,
         amplitude: float,
+        phase: float,
         t0: float,
         ramp: float = None,
         ramp_fraction: float = 0.1,
@@ -636,6 +829,7 @@ class CosineRampWaveform(BasicWaveform):
             ts: Time values at which to evaluate the pulse.
             width: The total width of the cosine ramp pulse, including ramp times.
             amplitude: The amplitude of the pulse.
+            phase: The phase of the pulse.
             t0: The starting time of the pulse.
             ramp: The ramp fraction. The pulse will be smoothly varied from 0 to
                 amplitude and vice versa over a time (width * ramp) at the
@@ -653,19 +847,19 @@ class CosineRampWaveform(BasicWaveform):
         else:
             raise ValueError("One of ramp or ramp_fraction must be specified!")
 
-        wave = np.zeros_like(ts, dtype=np.float32)
+        wave = np.zeros_like(ts, dtype=np.complex64)
+
+        rphi = amplitude * np.exp(1j * phase * np.pi / 180)
 
         ramp_up = (0 <= ts) & (ts < rlen)
-        wave[ramp_up] = 0.5 * amplitude * (1 - np.cos(np.pi / rlen * ts[ramp_up]))
+        wave[ramp_up] = 0.5 * rphi * (1 - np.cos(np.pi / rlen * ts[ramp_up]))
 
         const = (rlen <= ts) & (ts < width - rlen)
-        wave[const] = amplitude
+        wave[const] = rphi
 
         ramp_down = (width - rlen <= ts) & (ts < width)
         wave[ramp_down] = (
-            0.5
-            * amplitude
-            * (1 + np.cos(np.pi / rlen * (ts[ramp_down] - width + rlen)))
+            0.5 * rphi * (1 + np.cos(np.pi / rlen * (ts[ramp_down] - width + rlen)))
         )
 
         return wave
@@ -679,16 +873,24 @@ class DRAG(Waveform):
     lmbda2: NumberOrExpression = 0
 
     @property
-    def width(self) -> float:
+    def width(self) -> NumberOrExpression:
         return self.envelope.width
 
     @property
-    def t0(self) -> float:
+    def t0(self) -> NumberOrExpression:
         return self.envelope.t0
 
     @property
-    def amplitude(self) -> float:
+    def amplitude(self) -> NumberOrExpression:
         self.envelope.amplitude
+
+    @property
+    def phase(self) -> NumberOrExpression:
+        return self.envelope.phase
+
+    @property
+    def channel(self) -> str:
+        return self.envelope.channel
 
     def evaluate_timepoints(
         self, ts: np.ndarray, lmbda: float, lmbda2: float, **kwargs: Any
@@ -707,10 +909,14 @@ class DRAG(Waveform):
         """
         envelope = self.envelope(ts, **kwargs)
 
-        d1 = np.gradient(envelope)
-        d2 = np.gradient(d1)
+        d1 = np.gradient(envelope, ts)
+        d2 = np.gradient(d1, ts)
 
         return envelope + 1j * lmbda * d1 + lmbda2 * d2
+
+    def assign_channel(self, new_channel, /) -> Self:
+        """Returns a modified waveform with a new channel."""
+        return self.evolve(envelope_channel=new_channel)
 
 
 # ========== Waveform converters ========== #
@@ -739,6 +945,14 @@ def make_waveform_structure_fn(cls):
             val["frame"] = val["mod_key"]
             del val["mod_key"]
 
+        if "channels" in val:
+            logger.warning(
+                "Waveform 'channels' has been renamed to 'channel' and is now "
+                "deprecated. Waveforms can only have a single string channel now."
+            )
+            val["channel"] = "".join(sorted(val["channels"]))
+            del val["channels"]
+
         return qwip.converter.structure(val, subclass)
 
     return structure_fn
@@ -765,6 +979,7 @@ __all__ = [
     "register_waveform",
     "Waveform",
     "BasicWaveform",
+    "ConvolvedWaveform",
     "InfiniteWaveform",
     "Marker",
     "TriggeredWaveform",